fix: rename containers, images and services
