--- test/ping-pong/user-local1/.dockerignore
@@ -1,22 +1,0 @@
-# doc
-README.md
-
-#docker
-docker-compose.yml
-Dockerfile
-
-# git
-.git
-.gitattributes
-.gitignore
-.github
-
-# Node
-node_modules/
-
-#dev
-.prettierignore
-.prettierrc.js
-.releaserc.json
-.eslintrc
-.eslintignore
