--- test/ping-pong/user-local1/docker-entrypoint.sh
@@ -1,6 +1,0 @@
-#! /bin/bash
-set -o errexit
-set -o pipefail
-# wait until there's a optimist instance up
-while ! nc -z ${CLIENT_HOST:-client} ${CLIENT_PORT:-80}; do sleep 3; done
-exec "$@"
