--- test/ping-pong/user-local1/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "User",
-  "version": "1.0.0",
-  "description": "'makes a container with user functionality'",
-  "main": "index.js",
-  "scripts": {
-    "prestart": "./pre-start-script.sh",
-    "start": "node --trace-warnings src/index.mjs",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "westlad",
-  "license": "CC0-1.0",
-  "dependencies": {
-    "async-mutex": "^0.3.2",
-    "axios": "^0.24.0",
-    "web3": "^1.6.1",
-    "ws": "^8.2.3",
-    "winston": "^3.3.3",
-    "config": "^3.3.1",
-    "common-files": "file:../../../common-files"
-  }
-}
