--- test/ping-pong/user-local1/pre-start-script.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-DIR=../../../common-files/node_modules
-if [[ -d "$DIR" ]]; then
-  rm -dr ../../../common-files/node_modules
-fi
-
-rm ./node_modules/common-files
-cp -R ../../../common-files ./node_modules
